Extend the image localization system to sounds and music
