+++ src/game_config_manager.cpp
@@ -525,6 +525,7 @@ void game_config_manager::reload_changed_game_config()
 	init_game_config(FORCE_RELOAD);
 
 	image::flush_cache();
+	sound::flush_cache();
 }
 
 void game_config_manager::load_game_config_for_editor()
