+++ src/picture.cpp
@@ -17,15 +17,12 @@
  * Routines for images: load, scale, re-color, etc.
  */
 
-#define GETTEXT_DOMAIN "wesnoth-lib"
-
 #include "picture.hpp"
 
 #include "config.hpp"
 #include "display.hpp"
 #include "filesystem.hpp"
 #include "game_config.hpp"
-#include "gettext.hpp"
 #include "image_modifications.hpp"
 #include "log.hpp"
 #include "preferences/general.hpp"
@@ -463,91 +460,6 @@ bool locator::value::operator<(const value& a) const
 	return false;
 }
 
-// Check if localized file is up-to-date according to l10n track index.
-// Make sure only that the image is not explicitly recorded as fuzzy,
-// in order to be able to use non-tracked images (e.g. from UMC).
-static std::set<std::string> fuzzy_localized_files;
-static bool localized_file_uptodate(const std::string& loc_file)
-{
-	if(fuzzy_localized_files.empty()) {
-		// First call, parse track index to collect fuzzy files by path.
-		std::string fsep = "\xC2\xA6"; // UTF-8 for "broken bar"
-		std::string trackpath = filesystem::get_binary_file_location("", "l10n-track");
-
-		// l10n-track file not present. Assume image is up-to-date.
-		if(trackpath.empty()) {
-			return true;
-		}
-
-		std::string contents = filesystem::read_file(trackpath);
-
-		for(const std::string& line : utils::split(contents, '\n')) {
-			std::size_t p1 = line.find(fsep);
-			if(p1 == std::string::npos) {
-				continue;
-			}
-
-			std::string state = line.substr(0, p1);
-			boost::trim(state);
-			if(state == "fuzzy") {
-				std::size_t p2 = line.find(fsep, p1 + fsep.length());
-				if(p2 == std::string::npos) {
-					continue;
-				}
-
-				std::string relpath = line.substr(p1 + fsep.length(), p2 - p1 - fsep.length());
-				fuzzy_localized_files.insert(game_config::path + '/' + relpath);
-			}
-		}
-
-		fuzzy_localized_files.insert(""); // make sure not empty any more
-	}
-
-	return fuzzy_localized_files.count(loc_file) == 0;
-}
-
-// Return path to localized counterpart of the given file, if any, or empty string.
-// Localized counterpart may also be requested to have a suffix to base name.
-static std::string get_localized_path(const std::string& file, const std::string& suff = "")
-{
-	std::string dir = filesystem::directory_name(file);
-	std::string base = filesystem::base_name(file);
-
-	const std::size_t pos_ext = base.rfind(".");
-
-	std::string loc_base;
-	if(pos_ext != std::string::npos) {
-		loc_base = base.substr(0, pos_ext) + suff + base.substr(pos_ext);
-	} else {
-		loc_base = base + suff;
-	}
-
-	// TRANSLATORS: This is the language code which will be used
-	// to store and fetch localized non-textual resources, such as images,
-	// when they exist. Normally it is just the code of the PO file itself,
-	// e.g. "de" of de.po for German. But it can also be a comma-separated
-	// list of language codes by priority, when the localized resource
-	// found for first of those languages will be used. This is useful when
-	// two languages share sufficient commonality, that they can use each
-	// other's resources rather than duplicating them. For example,
-	// Swedish (sv) and Danish (da) are such, so Swedish translator could
-	// translate this message as "sv,da", while Danish as "da,sv".
-	std::vector<std::string> langs = utils::split(_("language code for localized resources^en_US"));
-
-	// In case even the original image is split into base and overlay,
-	// add en_US with lowest priority, since the message above will
-	// not have it when translated.
-	langs.push_back("en_US");
-	for(const std::string& lang : langs) {
-		std::string loc_file = dir + "/" + "l10n" + "/" + lang + "/" + loc_base;
-		if(filesystem::file_exists(loc_file) && localized_file_uptodate(loc_file)) {
-			return loc_file;
-		}
-	}
-
-	return "";
-}
-
 // Ensure PNG images with an indexed palette are converted to 32-bit RGBA.
 static void standardize_surface_format(surface& surf)
 {
@@ -582,7 +494,7 @@ static surface load_image_file(const image::locator& loc)
 	{
 		if(!location.empty()) {
 			// Check if there is a localized image.
-			const std::string loc_location = get_localized_path(location);
+			const std::string loc_location = filesystem::get_localized_path(location);
 			if(!loc_location.empty()) {
 				location = loc_location;
 			}
@@ -594,7 +506,7 @@ static surface load_image_file(const image::locator& loc)
 
 			// If there was no standalone localized image, check if there is an overlay.
 			if(!res.null() && loc_location.empty()) {
-				const std::string ovr_location = get_localized_path(location, "--overlay");
+				const std::string ovr_location = filesystem::get_localized_path(location, "--overlay");
 				if(!ovr_location.empty()) {
 					add_localized_overlay(ovr_location, res);
 				}
